rename grid components
